move "optimization" part to Processor.py
