+++ src/Processor.py
@@ -1,3 +1,6 @@
+import copy
+from datetime import date, timedelta
+
 import Content
 import re
 
@@ -37,12 +40,66 @@ def simple_process(arrangement):
     return new_arrangement
 
 
-class Processor:
-    def __init__(self):
-        # make column names configurable maybe
-        pass
+def _routine(arrangements, user_balance, recommended_result):
+    user_balance_cpy = copy.deepcopy(user_balance)
+    for user in user_balance:
+        if user_balance[user] == 0:
+            del user_balance_cpy[user]
+    user_balance = user_balance_cpy
 
-    def process(self, content: Content):
+    minimums = dict((k, v) for k, v in user_balance.items() if v < 0)
+    maximums = dict((k, v) for k, v in user_balance.items() if v > 0)
+
+    try:
+        max_minimum = max(minimums, key=minimums.get)
+        max_maximum = max(maximums, key=maximums.get)
+    except ValueError:
+        return
+
+    max_minimum_value = round(user_balance[max_minimum], 2)
+    max_maximum_value = round(user_balance[max_maximum], 2)
+
+    summation = max_minimum_value + max_maximum_value
+
+    amount = max_maximum_value
+    if summation > 0:
+        del user_balance[max_minimum]
+        user_balance[max_maximum] += max_minimum_value
+        # procedures.extend(build_transfer_base_on_transfer_order(arrangements[max_minimum], max_minimum, max_maximum, -max_minimim_value))
+        # print(f"{max_minimum} 转给 {max_maximum} {-max_minimim_value}刀")
+
+        amount = -max_minimum_value
+
+    elif summation < 0:
+        del user_balance[max_maximum]
+        user_balance[max_minimum] += max_maximum_value
+        # procedures.extend(build_transfer_base_on_transfer_order(arrangements[max_minimum], max_minimum, max_maximum, max_maximum_value))
+        # print(f"{max_minimum} 转给 {max_maximum} {max_maximum_value}刀")
+    elif summation == 0:
+        del user_balance[max_maximum]
+        del user_balance[max_minimum]
+        # procedures.extend(build_transfer_base_on_transfer_order(arrangements[max_minimum], max_minimum, max_maximum, max_maximum_value))
+        # print(f"{max_minimum} 转给 {max_maximum} {max_maximum_value}刀")
+
+    if amount != 0:
+        if max_minimum not in recommended_result:
+            recommended_result[max_minimum] = dict()
+        if max_maximum not in recommended_result[max_minimum]:
+            recommended_result[max_minimum][max_maximum] = dict()
+
+        recommended_result[max_minimum][max_maximum] = amount
+    _routine(arrangements, user_balance, recommended_result)
+
+
+class Processor:
+    def __init__(self, content: Content):
+        # make column names configurable maybe
+        self.content = content
+        self.user_statistics = None
+        self.result = None
+
+    def process(self):
+        content = self.content
         stat_columns = ["date", "user", "amount", "event_tag"]
         user_stat = dict()
         for column in stat_columns:
@@ -146,26 +203,94 @@ class Processor:
                 result[to_who][from_who] += how_much
 
         # 最后把它变成谁给谁转钱
-        real_final_result = dict()
+        final_result = dict()
         # init
         for user in users:
-            real_final_result[user] = dict()
+            final_result[user] = dict()
 
         for key in list(result.keys()):
             for user in list(result[key].keys()):
-                real_final_result[user][key] = result[key][user]
+                final_result[user][key] = result[key][user]
 
         # deep process 为了避免你给我转十块我给你转十块的事情发生
-        real_final_result = simple_process(real_final_result)
+        final_result = simple_process(final_result)
 
         # 去掉0
-        real_final_result = clean_zero_node(real_final_result)
+        final_result = clean_zero_node(final_result)
 
         # 那0都去掉了，就可能出现空的dict，也要去掉
-        for each in list(real_final_result.keys()):
-            if len(real_final_result[each]) == 0:
-                del real_final_result[each]
+        for each in list(final_result.keys()):
+            if len(final_result[each]) == 0:
+                del final_result[each]
 
         stat_df = pd.DataFrame.from_dict(user_stat)
+        self.user_statistics = stat_df
 
-        return real_final_result, stat_df
+        self.result = final_result
+        return final_result
+
+    def get_summary(self):
+        user_statistics = self.user_statistics
+        total_summary = dict()
+        current_month_summary = dict()
+        previouse_month_summary = dict()
+        event_summary = dict()
+
+        today = date.today()
+        curr_month_start = today.replace(day=1)
+        last_month_end = curr_month_start - timedelta(days=1)
+        last_month_start = last_month_end.replace(day=1)
+
+        users = user_statistics["user"].unique()
+        for user in users:
+            user_df = user_statistics.loc[user_statistics["user"] == user]
+            user_df_last_month = user_df.loc[
+                (user_df['date'] >= last_month_start) & (user_df['date'] <= last_month_end)]
+            user_df_curr_month = user_df.loc[(user_df['date'] >= curr_month_start) & (user_df['date'] <= today)]
+
+            total_summary[user] = user_df["amount"].sum()
+            current_month_summary[user] = user_df_curr_month["amount"].sum()
+            previouse_month_summary[user] = user_df_last_month["amount"].sum()
+
+            if total_summary[user] == 0:
+                del total_summary[user]
+            if current_month_summary[user] == 0:
+                del current_month_summary[user]
+            if previouse_month_summary[user] == 0:
+                del previouse_month_summary[user]
+
+            user_tags = user_df["event_tag"].unique().tolist()
+            user_tags = [str.strip(i) for i in user_tags]
+            if "" in user_tags:
+                user_tags.remove("")
+            for tag in user_tags:
+                if tag not in event_summary:
+                    event_summary[tag] = dict()
+                event_summary[tag][user] = user_df.loc[user_df["event_tag"] == tag]["amount"].sum()
+
+        return total_summary, current_month_summary, previouse_month_summary, event_summary
+    
+    def get_optimized(self):
+        if self.result is None:
+            self.process()
+        result_copy = copy.deepcopy(self.result)
+        df = self.content.get_df()
+        df["who"] = df["who"].apply(lambda x: str(x).replace("，", ",").replace("（", "(").replace("）", ")"))
+        users = self.content.get_users()
+        user_balance = dict()
+        for user in users:
+            user_balance[user] = 0
+
+        for user in result_copy:
+            for sub_user in result_copy[user]:
+                user_balance[user] -= result_copy[user][sub_user]
+                user_balance[sub_user] += result_copy[user][sub_user]
+
+        # round the user balance to 2 decimal places
+        for sub_user in user_balance:
+            user_balance[sub_user] = round(user_balance[sub_user], 2)
+
+        recommended_result = dict()
+        _routine(result_copy, user_balance, recommended_result)
+
+        return recommended_result
